+++ src/helpers/comparisonKey.ts
@@ -4,9 +4,5 @@ export default function comparisonKey<ElementType>(
   val: ElementType,
   toKey: ((a: ElementType) => string | number | bigint) | undefined = undefined
 ) {
-  return (val as unknown as Dream)?.isDreamInstance
-    ? (val as unknown as Dream).comparisonKey
-    : toKey
-      ? toKey(val)
-      : val
+  return val instanceof Dream ? val.comparisonKey : toKey ? toKey(val) : val
 }